test: added other_test and included regions which are shown in the console
